requirements.txt
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.5.0
streamlit==1.12.0

